chore(Pulumi.experiments.yaml): update image tags and commands for consistency and debugging
The image tags for the modware components have been updated to the
latest versions to ensure that the most recent features and fixes are
included. The command for the load-content-from-s3 service has been
changed to 'dataloader' for clarity, and a logLevel has been added to
the modware-stock configuration to facilitate debugging. Additionally,
the port for modware-stock has been changed to avoid conflicts with
other services.