feat(configMap.go): add CreateConfigMap method to dynamically create a ConfigMap based on provided configuration values
The CreateConfigMap method is added to the GraphqlServer struct to dynamically generate a ConfigMap object in Kubernetes. This method takes in a Pulumi context and uses the configuration values provided to populate the data field of the ConfigMap. This allows for more flexibility and ease of configuration management when deploying the GraphQL server.