fix(Pulumi.experiments.yaml): update image tags to latest commit hashes for frontpage, publication, and stock-center apps
The image tags for the frontpage, publication, and stock-center apps have been updated to the latest commit hashes (develop-90ffc7c) to ensure that the correct versions of the images are being used in the Pulumi experiments configuration. This update helps in maintaining consistency and tracking changes in the deployed applications.