docs(api, sa): add documentation for enabling, disabling APIs and creating, granting roles to a service account
The commit adds documentation for enabling and disabling APIs in the Google Cloud Project, as well as creating a service account, granting the Cloud Manager role to the service account, and generating a service key. This documentation will help users understand the necessary steps to manage APIs and service accounts effectively in the project.