feat(roles_permissions.go): add new function addRoles to manage IAM roles for a service account
The new function addRoles is added to the roles_permissions.go file to handle the addition of IAM roles for a specific service account. This function allows for the management of IAM roles by providing the project name, service account name, and the roles to be added. It utilizes the IAM service to interact with IAM policies and bindings for the specified service account.