feat(service_account.go): add functionality to read roles from a file and assign them to a service account
The import statements in the file have been reformatted to improve readability. Additionally, a new function `readRolesFromFile` has been added to read roles from a file specified by the `filePath` parameter. These roles are then assigned to a service account using the `addRolesToServiceAccount` method. This new functionality enhances the flexibility of the service account creation process by allowing roles to be specified in a file.
issue #136