feat(service_account.go): add functionality to create a service account key
The function createServiceAccountKey() has been added to the IAMClient struct in service_account.go. This function allows for the creation of a service account key by making a request to the Google Cloud IAM API. The key is then written to a file specified by the outputPath parameter. This new functionality enhances the ability to manage service accounts within the application.
issue #136