feat(README.md): add documentation for setting up a Kubernetes cluster on GCP
This commit adds documentation for setting up a Kubernetes cluster on Google Cloud Platform for dictyBase applications. It includes instructions on installing prerequisites such as asdf, just, and direnv, as well as acquiring the Service Account Manager key and enabling required APIs. The documentation provides a step-by-step guide for users to follow in order to set up their cluster.