docs(README.md): update README with kops-cluster-creator setup instructions
The README is updated to include instructions on how to set up the
`kops-cluster-creator` service account and update the
`GOOGLE_APPLICATION_CREDENTIALS` environment variable. This ensures
that the Go Google Cloud client libraries use the
`kops-cluster-creator` service key for managing resources on the
Google Cloud Platform project. Also added a command to load env vars.