feat(Justfile): add recipe to set environment variables
Adds a new recipe `set-env-var` to the Justfile that uses a go
binary to set environment variables. This recipe takes two
arguments: `name` and `value`. It builds a go binary, then executes
it to set the environment variable. Finally, it calls `direnv allow`
to update the environment and prints a message to the console.

This change allows developers to easily set environment variables
during development and deployment.