feat(README.md): add single command cluster setup instruction
The commit adds `direnv` to the list of installed binaries, as it is a prerequisite for setting up the environment. It also adds a new section that describes how to set up the cluster with a single command using `just init-kops-cluster`. The cluster setup steps are reordered to improve the flow of the documentation. The link to the ADC setup documentation is updated to point to the correct section.