feat(dcr-configuration): add support for creating Kubernetes secrets
This commit introduces the basic structure for configuring DCR (Dynamic Client Registration) using Pulumi. It includes functionality to read configuration from Pulumi, create Kubernetes secrets based on the configuration, and export the names of the created secrets. The changes allow for a declarative approach to managing DCR configurations within a Kubernetes environment.