docs(README.md): add documentation for deploying applications using Pulumi
This commit adds documentation to the README.md file that describes how to deploy applications to the Kubernetes cluster using Pulumi. It includes instructions for creating a Pulumi Manager Service Account Key, setting the PULUMI_GCP_CREDENTIALS environment variable, initializing the Pulumi State Store, and performing initial deployments. It also lists the Pulumi projects that need to be created first because they create secrets that other projects rely on.