feat(pulumi.justfile): add recipe to create initial resources
This commit introduces a new recipe `create-initial-resources` to
automate the creation of stacks and resources based on a list of
projects defined in `initial-resources.txt`. It also introduces a
recipe to create stacks from existing stacks. This enhancement
streamlines the setup process for new projects by allowing the
initial infrastructure to be provisioned automatically.