feat(Justfile): add pulumi-init-and-deploy recipe to create initial resources
The `setup-pulumi-deployment` recipe is renamed to `initialize-pulumi` to better reflect its purpose. The `pulumi-init-and-deploy` recipe is added to initialize Pulumi and create initial resources. The `create-initial-resources` recipe is removed as its functionality is now incorporated into the `pulumi-init-and-deploy` recipe. This change simplifies the process of setting up and deploying Pulumi resources.